backend/requirements.txt
fastapi
uvicorn
python-docx
langgraph
langchain-openai
python-multipart
pydantic
rich
colorama
spotipy
gtts
pydub
RestrictedPython
python-dotenv
pygame
requests
speechrecognition
scikit-learn
boto3
langchain
psycopg2
simpleaudio
sentence-transformers
chromadb
docx
pyttsx3
pyaudio
soundfile
torch
openai-whisper
asyncpg